Feat(nix): Add flake and direnv config
The `flake.nix` file defines two handy outputs:
1. The shell environment that comes the Rust toolchain,
2. The `git` pre-commit hooks that check for Nix and Rust formatting.

The `.envrc` is only here to automatically install and setup all the
above when I enter the directory.

This is absolutely not necessary for the project to run on your side but
I don't have the Rust toolchain installed on my work machine so this
allows me to compile your project without altering my system.